+++ MapboxCoreNavigation/RouteProgress.swift
@@ -406,6 +406,15 @@ open class RouteStepProgress: NSObject {
      */
     @objc public var spokenInstructionIndex: Int = 0
     
+    
+    /**
+     An `Array` of remaining `SpokenInstruction` for a step.
+     */
+    @objc public var remainingSpokenInstructions: [SpokenInstruction]? {
+        guard let instructions = step.instructionsSpokenAlongStep else { return nil }
+        return Array(instructions.suffix(from: spokenInstructionIndex))
+    }
+    
     /**
      Current Instruction for the user's progress along a step.
      */